+++ CHANGELOG.md
@@ -328,6 +328,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Core\ShopControl::_getFrontendStartController()`, Use getFrontendStartControllerKey()
 - `\OxidEsales\EshopCommunity\Core\WidgetControl::$_blHandlerSet`
 - `\OxidEsales\EshopCommunity\Core\ViewConfig::getSession()`
+- `\OxidEsales\EshopCommunity\Core\Routing\Module\ClassProviderStorage`
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)